feat(rgb): added the "parseRGB" function
The "parseRGB" function tries to parse the given input string as CSS RGB
color. It throws an InvalidValueError when the input is invalid or one
of the values is out of range (0-255).
It returns a "ParsedRGB" object when the parsing was successful.
Also added the tests for the "parseRGB" function and added the reexport
from the `index.ts` file for easier usage.